updated calculator
based on following feedback:
Layout constraints break on changing orientation
Changing orientation starts the buttons rearranging from the top left corner. Instead, it should re-adjust the buttons as per the orientation.
There is a better approach than creating and adding a set of buttons every time user changes the orientation. We can have a discussion over this.
The ViewController class contains the logic / brain of the Calculator. A good and recommended practice is to extract the code into a Model / ViewModel or a Singleton class as per the situation.